New navbar approach, now its own component, add to pages as needed. Variable dropdown, WIP
